Update TESTING.R
e Internationalidditch Federation (IQF) has contracted you to assist in their analysis and reporting of
results. You have been provided the data le competitionResults.csv containing details of each match
(the teams that played, which team caught the snitch, and each team’s number of goals).
e IQF wants you to use the le to provide the needed four pieces of analysis below. e IQF also
want to use your code for analysis of future data les.
1. Write the code to compute the total scored by each team for each match. Note that each goal is
worth 10 points and the team that caught the snitch is awarded an extra 50 points. Run your code
on the data and present the results in a table showing the season and round number, the match
number (a new variable that identies each match of the season), the home and away team names
and their scores
2. e winner of each match is the team who has the highest score. Each team scores 3 points for a
win and 1 point for a draw and the team with the most points at the end of the season is the season
winner. Write the code to compute each team’s points aer the rst 5 rounds of matches and the
full season. Provide the results for a season in a table (two tables, one table for each set of results),
ordering teams by their points (i.e. a results ladder). Pick any season.
3. An investigation into home game advantage is planned for the coming months. Write a function
that returns a count of the number of home games each team has had, the total number of games
each team has played in, the total points scored in a home game and the total number of points the
team has scored. Provide these results in a data frame. Run the function on the provided data and
show the resulting table output.
4. Provide a table containing the team statistics for all seasons. Including, but not limited to the following: number of games, number of wins, winning percentage, points scored, points conceded,
number of snitch catches, goals scored, number of tournament wins. Order the teams in terms of
total points scored